refactor(market): renamed DemandOfferEvent to MarketProposalEvent
