+++ src/market/agreement/index.ts
@@ -1,3 +1,4 @@
 export { Agreement, LegacyAgreementServiceOptions, ProviderInfo, AgreementState } from "./agreement";
 export { AgreementCandidate, AgreementSelector } from "./selector";
 export { MarketApiConfig } from "./config";
+export * from "./agreement-event";
