+++ src/market/demand/index.ts
@@ -1,3 +1,2 @@
 export * from "./demand";
-export * from "./demand-events";
 export * from "./demand-body-builder";
