+++ src/market/proposal/index.ts
@@ -1,0 +1,7 @@
+export * from "./types";
+export * from "./market-proposal-event";
+export * from "./market-proposal";
+export * from "./offer-proposal";
+export * from "./offer-counter-proposal";
+export * from "./proposals_batch";
+export * from "./proposal-properties";
